requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
distro==1.9.0
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
groq==0.18.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.34
langchain-groq==0.2.4
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.6
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.15
packaging==24.2
preshed==3.0.9
propcache==0.2.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
tenacity==9.0.0
thinc==8.3.4
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0

# Added for user authentication and database
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
psycopg2-binary==2.9.9
email-validator==2.2.0
pytest

# Added for Cloudinary image handling
cloudinary==1.41.0

# Added for Google OAuth
Authlib==1.3.0
requests==2.32.3
requests-oauthlib==1.3.1

# Added for RAG-based movie recommendations
chromadb==0.4.22
sentence-transformers==2.3.1
huggingface-hub==0.20.3
